Save pages as plain HTML to enable vanilla Firefox
Save pages as plain HTML and use `saveAs: false`.

This ensures the extension will always save files
to the MemoryCache subdirectory even when the user
enables the "Always ask you where to save files" preference.

I could not find a way to make this work for PDFs.
The tabs.saveAsPDF WebExtension API does not seem to
have an equivalent to `saveAs: false`, which is why
Liv was building a custom version of Firefox.